fix(fs.helpers.ts): ensure folder exists before cleaning up

+++ packages/system-api/src/modules/fs/fs.helpers.ts
@@ -43,12 +43,12 @@ export const getSeed = () => {
 };
 
 export const ensureAppFolder = (appName: string, cleanup = false) => {
-  if (cleanup) {
+  if (cleanup && fileExists(`/apps/${appName}`)) {
     deleteFolder(`/apps/${appName}`);
   }
 
   if (!fileExists(`/apps/${appName}/docker-compose.yml`)) {
-    deleteFolder(`/apps/${appName}`);
+    if (fileExists(`/apps/${appName}`)) deleteFolder(`/apps/${appName}`);
     // Copy from apps repo
     fs.copySync(getAbsolutePath(`/repos/${config.APPS_REPO_ID}/apps/${appName}`), getAbsolutePath(`/apps/${appName}`));
   }
